+++ Libs/libMacGitverCore/RepoMan/Frontend/Namespace.cpp
@@ -43,62 +43,4 @@ namespace RM
         return d->name;
     }
 
-    //-- NamespacePrivate --------------------------------------------------------------------------
-
-    NamespacePrivate::NamespacePrivate(Namespace* _pub, const QString& _name)
-        : BasePrivate(_pub)
-        , name(_name)
-    {
-    }
-
-    ObjTypes NamespacePrivate::objType() const
-    {
-        return ObjTypes::Namespace;
-    }
-
-    void NamespacePrivate::dumpSelf(Internal::Dumper& dumper) const
-    {
-        dumper.addLine(QString(QStringLiteral("Namespace 0x%1 - %2"))
-                       .arg(quintptr(mPub),0,16)
-                       .arg(name));
-    }
-
-    void NamespacePrivate::postCreation()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->namespaceCreated(repository(), pub<Namespace>());
-        }
-
-        BasePrivate::postCreation();
-    }
-
-    void NamespacePrivate::preTerminate()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->namespaceAboutToBeDeleted(repository(), pub<Namespace>());
-        }
-
-        BasePrivate::preTerminate();
-    }
-
-    bool NamespacePrivate::refreshSelf()
-    {
-        return true;
-    }
-
-    QString NamespacePrivate::displayName() const
-    {
-        return name;
-    }
-
-    QString NamespacePrivate::objectTypeName() const
-    {
-        return QStringLiteral("Namespace");
-    }
-
-    bool NamespacePrivate::inherits(ObjTypes type) const
-    {
-        return type == ObjTypes::Namespace || BasePrivate::inherits(type);
-    }
-
 }
