+++ Libs/libMacGitverCore/RepoMan/Frontend/RefTreeNode.cpp
@@ -44,68 +44,4 @@ namespace RM
         return d->name;
     }
 
-    //-- RefTreeNodePrivate ------------------------------------------------------------------------
-
-    RefTreeNodePrivate::RefTreeNodePrivate(RefTreeNode* _pub, const QString& _name)
-        : BasePrivate(_pub)
-        , name(_name)
-    {
-    }
-
-    ObjTypes RefTreeNodePrivate::objType() const
-    {
-        return ObjTypes::RefTreeNode;
-    }
-
-    void RefTreeNodePrivate::dumpSelf(Dumper& dumper) const
-    {
-        dumper.addLine(QString(QStringLiteral("RefTreeNode 0x%1 [%2]"))
-                       .arg(quintptr(mPub),0,16)
-                       .arg(name));
-    }
-
-    void RefTreeNodePrivate::postCreation()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->refTreeNodeCreated(repository(), pub<RefTreeNode>());
-        }
-
-        BasePrivate::postCreation();
-    }
-
-    void RefTreeNodePrivate::preTerminate()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->refTreeNodeAboutToBeDeleted(repository(), pub<RefTreeNode>());
-        }
-
-        BasePrivate::preTerminate();
-    }
-
-    QString RefTreeNodePrivate::displayName() const
-    {
-        return name;
-    }
-
-    bool RefTreeNodePrivate::refreshSelf()
-    {
-        // We don't have anything to refresh - we're purely virutual
-        return true;
-    }
-
-    bool RefTreeNodePrivate::refreshCheckDispensable()
-    {
-        return mChildren.isEmpty();
-    }
-
-    QString RefTreeNodePrivate::objectTypeName() const
-    {
-        return QStringLiteral("RefTreeNode");
-    }
-
-    bool RefTreeNodePrivate::inherits(ObjTypes type) const
-    {
-        return type == ObjTypes::RefTreeNode || BasePrivate::inherits(type);
-    }
-
 }
