+++ Libs/libMacGitverCore/RepoMan/Frontend/Remote.cpp
@@ -54,63 +54,5 @@ namespace RM
         return d->name;
     }
 
-    //-- RemotePrivate -----------------------------------------------------------------------------
-
-    RemotePrivate::RemotePrivate(Remote* _pub, const Git::Remote& _obj)
-        : BasePrivate(_pub)
-    {
-        name = _obj.name();
-    }
-
-    ObjTypes RemotePrivate::objType() const
-    {
-        return ObjTypes::Remote;
-    }
-
-    void RemotePrivate::dumpSelf(Internal::Dumper& dumper) const
-    {
-        dumper.addLine(QString(QStringLiteral("Remote %2 0x%1"))
-                       .arg(quintptr(mPub),0,16)
-                       .arg(name));
-    }
-
-    void RemotePrivate::postCreation()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->remoteCreated(repository(), pub<Remote>());
-        }
-
-        BasePrivate::postCreation();
-    }
-
-    void RemotePrivate::preTerminate()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->remoteAboutToBeDeleted(repository(), pub<Remote>());
-        }
-
-        BasePrivate::preTerminate();
-    }
-
-    QString RemotePrivate::displayName() const
-    {
-        return name;
-    }
-
-
-    bool RemotePrivate::refreshSelf()
-    {
-        return true;
-    }
-
-    QString RemotePrivate::objectTypeName() const
-    {
-        return QStringLiteral("Remote");
-    }
-
-    bool RemotePrivate::inherits(ObjTypes type) const
-    {
-        return type == ObjTypes::Remote || BasePrivate::inherits(type);
-    }
-
 }
+
