+++ Libs/libMacGitverCore/RepoMan/Frontend/Submodule.cpp
@@ -43,50 +43,4 @@ namespace RM
         d->mIsInitializing = false;
     }
 
-    //-- SubmodulePrivate --------------------------------------------------------------------------
-
-    SubmodulePrivate::SubmodulePrivate(Submodule* pub, const Git::Repository& repo)
-        : RepoPrivate( pub, repo )
-    {
-    }
-
-    ObjTypes SubmodulePrivate::objType() const
-    {
-        return ObjTypes::Submodule;
-    }
-
-    void SubmodulePrivate::dumpSelf(Dumper& dumper) const
-    {
-        dumper.addLine(QString(QStringLiteral("Submodule 0x%1"))
-                       .arg(quintptr(mPub),0,16));
-    }
-
-    void SubmodulePrivate::postCreation()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->submoduleCreated(repository(), pub<Submodule>());
-        }
-
-        RepoPrivate::postCreation();
-    }
-
-    void SubmodulePrivate::preTerminate()
-    {
-        if (!repoEventsBlocked()) {
-            Events::self()->submoduleAboutToBeDeleted(repository(), pub<Submodule>());
-        }
-
-        RepoPrivate::preTerminate();
-    }
-
-    QString SubmodulePrivate::objectTypeName() const
-    {
-        return QStringLiteral("Submodule");
-    }
-
-    bool SubmodulePrivate::inherits(ObjTypes type) const
-    {
-        return type == ObjTypes::Submodule || RepoPrivate::inherits(type);
-    }
-
 }
